+++ src/main/java/br/com/fiap/myassist/repository/ObservacaoRepository.java
@@ -1,9 +1,8 @@
 package br.com.fiap.myassist.repository;
 
-import br.com.fiap.myassist.entity.Cliente;
 import br.com.fiap.myassist.entity.Observacao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ObservacaoRepository extends JpaRepository<Observacao, Long> {
+public interface ObservacaoRepository extends JpaRepository<Observacao, Integer> {
 
 }
